Keep Going Password Checker
This program demonstrates a loop with multiple exits.

Backstory:
So far, I've stuck the the coding challenges in the course I am taking. Today I was watching a seminar video on YouTube, and this question came up.
• Consider a basic password loop
• It exits with a positive result if the user chooses the right password
• It exits with a negative result if the user is wrong 3 times
• How will you code this loop?
I decided to pause the video and give it a try. I was able to complete the problem and it was super motivating.