Coffee Machine Program
The most challenging program to date. The machine should prompt the user for their order. 
If you type "off", it will turn the machine off.
If you type report, it will give you an inventory or the milk, water, coffee and profit.
When you type the drink you want, it will check the ingredients against the inventory.
If there is enough, it will make your drink, give change if necessary (or refund you if you don't give enough money), and add the profit to the report.
If there isn't enough inventory, it will ask you to pick another drink.