Turtle Race
Using for loops to change object state and create multiple instances of an object. The final product creates 6 different colored turtles that will each move forward by a random number of pixels until the first one reaches the edge of the screen. User is asked in the beginning to guess which turtle will win and then is told whether they won or not at the end of the race.